requirements.txt
Flask
Flask-PyMongo
python-dotenv
requests